EVG-16987 Move shared pod / host comm functions to base
