+++ config.go
@@ -36,7 +36,7 @@ var (
 	ClientVersion = "2022-05-26"
 
 	// Agent version to control agent rollover.
-	AgentVersion = "2022-05-24"
+	AgentVersion = "2022-06-08"
 )
 
 // ConfigSection defines a sub-document in the evergreen config
