+++ rest/route/agent_test.go
@@ -1,0 +1,86 @@
+package route
+
+import (
+	"context"
+	"net/http"
+	"testing"
+
+	"github.com/evergreen-ci/evergreen"
+	"github.com/evergreen-ci/evergreen/apimodels"
+	"github.com/mongodb/grip/send"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestAgentCedarConfig(t *testing.T) {
+	for tName, tCase := range map[string]func(ctx context.Context, t *testing.T, rh *agentCedarConfig, s *evergreen.Settings){
+		"FactorySucceeds": func(ctx context.Context, t *testing.T, rh *agentCedarConfig, s *evergreen.Settings) {
+			copied := rh.Factory()
+			assert.NotZero(t, copied)
+			_, ok := copied.(*agentCedarConfig)
+			assert.True(t, ok)
+		},
+		"ParseSucceeds": func(ctx context.Context, t *testing.T, rh *agentCedarConfig, s *evergreen.Settings) {
+			req, err := http.NewRequest(http.MethodGet, "https://example.com/rest/v2/agent/cedar_config", nil)
+			require.NoError(t, err)
+			assert.NoError(t, rh.Parse(ctx, req))
+		},
+		"RunSucceeds": func(ctx context.Context, t *testing.T, rh *agentCedarConfig, s *evergreen.Settings) {
+			resp := rh.Run(ctx)
+			require.NotZero(t, resp)
+			assert.Equal(t, http.StatusOK, resp.Status())
+
+			data, ok := resp.Data().(apimodels.CedarConfig)
+			require.True(t, ok)
+			assert.Equal(t, data.BaseURL, s.Cedar.BaseURL)
+			assert.Equal(t, data.RPCPort, s.Cedar.RPCPort)
+			assert.Equal(t, data.Username, s.Cedar.User)
+			assert.Equal(t, data.APIKey, s.Cedar.APIKey)
+		},
+		"ReturnsEmpty": func(ctx context.Context, t *testing.T, rh *agentCedarConfig, s *evergreen.Settings) {
+			*s = evergreen.Settings{}
+			resp := rh.Run(ctx)
+			require.NotZero(t, resp)
+			assert.Equal(t, http.StatusOK, resp.Status())
+
+			data, ok := resp.Data().(apimodels.CedarConfig)
+			require.True(t, ok)
+			assert.Zero(t, data)
+		},
+	} {
+		t.Run(tName, func(t *testing.T) {
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			s := &evergreen.Settings{
+				Splunk: send.SplunkConnectionInfo{
+					ServerURL: "server_url",
+					Token:     "token",
+					Channel:   "channel",
+				},
+				Providers: evergreen.CloudProviders{
+					AWS: evergreen.AWSConfig{
+						S3: evergreen.S3Credentials{
+							Bucket: "bucket",
+							Key:    "key",
+							Secret: "secret",
+						},
+						TaskSync: evergreen.S3Credentials{
+							Bucket: "bucket",
+							Key:    "key",
+							Secret: "secret",
+						},
+					},
+				},
+				LoggerConfig: evergreen.LoggerConfig{
+					LogkeeperURL: "logkeeper_url",
+				},
+			}
+
+			r, ok := makeAgentCedarConfig(s).(*agentCedarConfig)
+			require.True(t, ok)
+
+			tCase(ctx, t, r, s)
+		})
+	}
+}
