+++ rest/route/pod_agent_test.go
@@ -224,79 +224,6 @@ func TestPodAgentSetup(t *testing.T) {
 	}
 }
 
-func TestPodAgentCedarConfig(t *testing.T) {
-	for tName, tCase := range map[string]func(ctx context.Context, t *testing.T, rh *podAgentCedarConfig, s *evergreen.Settings){
-		"FactorySucceeds": func(ctx context.Context, t *testing.T, rh *podAgentCedarConfig, s *evergreen.Settings) {
-			copied := rh.Factory()
-			assert.NotZero(t, copied)
-			_, ok := copied.(*podAgentCedarConfig)
-			assert.True(t, ok)
-		},
-		"ParseSucceeds": func(ctx context.Context, t *testing.T, rh *podAgentCedarConfig, s *evergreen.Settings) {
-			req, err := http.NewRequest(http.MethodGet, "https://example.com/rest/v2/pod/agent/cedar_config", nil)
-			require.NoError(t, err)
-			assert.NoError(t, rh.Parse(ctx, req))
-		},
-		"RunSucceeds": func(ctx context.Context, t *testing.T, rh *podAgentCedarConfig, s *evergreen.Settings) {
-			resp := rh.Run(ctx)
-			require.NotZero(t, resp)
-			assert.Equal(t, http.StatusOK, resp.Status())
-
-			data, ok := resp.Data().(apimodels.CedarConfig)
-			require.True(t, ok)
-			assert.Equal(t, data.BaseURL, s.Cedar.BaseURL)
-			assert.Equal(t, data.RPCPort, s.Cedar.RPCPort)
-			assert.Equal(t, data.Username, s.Cedar.User)
-			assert.Equal(t, data.APIKey, s.Cedar.APIKey)
-		},
-		"ReturnsEmpty": func(ctx context.Context, t *testing.T, rh *podAgentCedarConfig, s *evergreen.Settings) {
-			*s = evergreen.Settings{}
-			resp := rh.Run(ctx)
-			require.NotZero(t, resp)
-			assert.Equal(t, http.StatusOK, resp.Status())
-
-			data, ok := resp.Data().(apimodels.CedarConfig)
-			require.True(t, ok)
-			assert.Zero(t, data)
-		},
-	} {
-		t.Run(tName, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			s := &evergreen.Settings{
-				Splunk: send.SplunkConnectionInfo{
-					ServerURL: "server_url",
-					Token:     "token",
-					Channel:   "channel",
-				},
-				Providers: evergreen.CloudProviders{
-					AWS: evergreen.AWSConfig{
-						S3: evergreen.S3Credentials{
-							Bucket: "bucket",
-							Key:    "key",
-							Secret: "secret",
-						},
-						TaskSync: evergreen.S3Credentials{
-							Bucket: "bucket",
-							Key:    "key",
-							Secret: "secret",
-						},
-					},
-				},
-				LoggerConfig: evergreen.LoggerConfig{
-					LogkeeperURL: "logkeeper_url",
-				},
-			}
-
-			r, ok := makePodAgentCedarConfig(s).(*podAgentCedarConfig)
-			require.True(t, ok)
-
-			tCase(ctx, t, r, s)
-		})
-	}
-}
-
 func TestPodAgentNextTask(t *testing.T) {
 	defer func() {
 		assert.NoError(t, db.ClearCollections(task.Collection, pod.Collection, dispatcher.Collection, event.AllLogCollection))
